requirements.txt
Django==3.1.2
markdown==3.2.2
readtime==1.1.1
mat2==0.11.0
django-upload-validator==1.1.5
django-clamd==0.4.0
django-ratelimit==3.0.1
